Feat/updates

See merge request farmz2u/african-highland-farmer!2

+++ Assets/Scripts/Livestock.cs
@@ -7,7 +7,7 @@ using UnityEngine.UI;
 public class Livestock : MonoBehaviour
 {
     public string type;
-    public int cost;
+    public double cost;
     public int maintenance_cost;
     public double manure_output;
     public int required_labor;
